Change vectors to arrays which makes everything ugly

+++ cell.cpp
@@ -1,11 +1,11 @@
 #include "cell.h"
 
-Cell::Cell(size_t i, size_t j) {
+Cell::Cell(int i, int j) {
     row = i;
     col = j;
 }
 
-Cell::Cell(size_t i, size_t j, int value)
+Cell::Cell(int i, int j, int value)
 {
     row = i;
     col = j;
@@ -39,22 +39,22 @@ char Cell::status_to_char() {
     }
 }
 
-size_t Cell::get_right(size_t ncols) {
+int Cell::get_right(int ncols) {
     if (col == ncols-1) return 0;
     else return col + 1;
 }
 
-size_t Cell::get_left(size_t ncols) {
+int Cell::get_left(int ncols) {
     if (col == 0) return ncols - 1;
     else return col - 1;
 }
 
-size_t Cell::get_top(size_t nrows) {
+int Cell::get_top(int nrows) {
     if (row == 0) return nrows - 1;
     else return row - 1;
 }
 
-size_t Cell::get_bottom(size_t nrows) {
+int Cell::get_bottom(int nrows) {
     if (row == nrows-1) return 0;
     else return row + 1;
 }
